refactor(support-level): move class to model module
