+++ packages/manager/apps/dedicated/client/app/account/user/support-level/support-level.constants.js
@@ -1,32 +1,5 @@
 export const API_MODEL_SUPPORT_LEVEL = 'me.SupportLevel.LevelTypeEnum';
 
-export const INACTIVE_SUPPORT_LEVEL = ['premium-accredited'];
-
-export const AVAILABLE_SUPPORT_LEVEL = ['business', 'enterprise'];
-
-export const FULL_SUPPORT_LEVEL_SUBSIDIARIES = [
-  'ASIA',
-  'AU',
-  'CA',
-  'CZ',
-  'DE',
-  'ES',
-  'FI',
-  'FR',
-  'GB',
-  'IE',
-  'IT',
-  'LT',
-  'NL',
-  'PL',
-  'PT',
-  'QC',
-  'SG',
-  'SN',
-  'WE',
-  'WS',
-];
-
 export const URLS = {
   ASIA: {
     standard: 'https://www.ovhcloud.com/asia/support-levels/standard/',
@@ -170,8 +143,5 @@ export const URLS = {
 
 export default {
   API_MODEL_SUPPORT_LEVEL,
-  AVAILABLE_SUPPORT_LEVEL,
-  FULL_SUPPORT_LEVEL_SUBSIDIARIES,
-  INACTIVE_SUPPORT_LEVEL,
   URLS,
 };
