+++ packages/manager/modules/models/support/level/support-level.constants.js
@@ -1,0 +1,32 @@
+export const INACTIVE_SUPPORT_LEVEL = ['premium-accredited'];
+
+export const AVAILABLE_SUPPORT_LEVEL = ['business', 'enterprise'];
+
+export const FULL_SUPPORT_LEVEL_SUBSIDIARIES = [
+  'ASIA',
+  'AU',
+  'CA',
+  'CZ',
+  'DE',
+  'ES',
+  'FI',
+  'FR',
+  'GB',
+  'IE',
+  'IT',
+  'LT',
+  'NL',
+  'PL',
+  'PT',
+  'QC',
+  'SG',
+  'SN',
+  'WE',
+  'WS',
+];
+
+export default {
+  AVAILABLE_SUPPORT_LEVEL,
+  FULL_SUPPORT_LEVEL_SUBSIDIARIES,
+  INACTIVE_SUPPORT_LEVEL,
+};
